requirements.txt
pandas==1.2.4
requests==2.25.1
python-dotenv==0.19.1
